20231022/python/pyproject.toml
[tool.poetry]
name = "kafka-demo"
version = "0.1.0"
description = ""
authors = ["Ashal aka JOKER <ahastudio@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.5"
kafka-python = "^2.0.2"
fastavro = "^1.8.4"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.10.0"
ruff = "^0.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79

[tool.ruff]
select = ["A", "B", "C"]
